Move datacodec into event
