+++ pkg/event/datacodec/doc.go
