+++ pkg/event/datacodec/json/data.go
