+++ pkg/event/datacodec/json/doc.go
