+++ pkg/event/datacodec/text/text.go
